テストヘルパーを切り出し
